+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to
 - cosmwasm-std: Replace `ContractInfoResponse::new` with new (unstable)
   constructor, remove `SubMsgExecutionResponse` (Use `SubMsgResponse` instead)
   and remove `PartialEq<&str> for Addr` (validate the address and use
-  `PartialEq<Addr>` instead). ([#1879])
+  `PartialEq<Addr> for Addr` instead). ([#1879])
 
 [#1879]: https://github.com/CosmWasm/cosmwasm/pull/1879