+++ CHANGELOG.md
@@ -19,11 +19,15 @@ and this project adheres to
   ([#1874])
 - cosmwasm-vm: Make `CacheOptions` non-exhaustive and add a constructor.
   ([#1898])
+- cosmwasm-std: `Coin::new` now takes `Into<Uint128>` instead of `u128` as the
+  first argument and `DecCoin::new` takes `Into<Decimal256>` instead of
+  `Decimal256`. ([#1902])
 
 [#1874]: https://github.com/CosmWasm/cosmwasm/pull/1874
 [#1879]: https://github.com/CosmWasm/cosmwasm/pull/1879
 [#1890]: https://github.com/CosmWasm/cosmwasm/pull/1890
 [#1898]: https://github.com/CosmWasm/cosmwasm/pull/1898
+[#1902]: https://github.com/CosmWasm/cosmwasm/pull/1902
 
 ## [1.5.0-rc.0]